manual: make the code example executable
All code examples should actually be executable.
It is clear to an experienced assembly programmer
that the former example was just meant as an
example but a less experienced person might
just copy and paste that example and experience
a failire because of the string constant
being executed as code.